third_party/blink/web_tests/fast/shapes/parsing/parsing-shape-image-threshold-expected.txt

Test parsing of the CSS shape-image-threshold property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyleValue("shape-image-threshold", "0") is "0"
PASS getComputedStyleValue("shape-image-threshold", "0.5") is "0.5"
PASS getComputedStyleValue("shape-image-threshold", "1") is "1"
PASS getComputedStyleValue("shape-image-threshold", "-0.1") is "0"
PASS getComputedStyleValue("shape-image-threshold", "1.1") is "1"
PASS getComputedStyleValue("shape-image-threshold", "identifier") is "0"
PASS getComputedStyleValue("shape-image-threshold", "'string'") is "0"
PASS getChildComputedStyle("shape-image-threshold", "0", "0") is "0"
PASS getChildComputedStyle("shape-image-threshold", "0", "1") is "1"
PASS getChildComputedStyle("shape-image-threshold", "1", "-1") is "0"
PASS getChildComputedStyle("shape-image-threshold", "2", "1") is "1"
PASS CSS.supports('shape-image-threshold', '0.8') is true
PASS CSS.supports('shape-image-threshold', '-0.5') is true
PASS CSS.supports('shape-image-threshold', '-34') is true
PASS CSS.supports('shape-image-threshold', '5px') is false
PASS successfullyParsed is true

TEST COMPLETE
